Reference

ConvolutionalRBMs.BinaryConvRBM — Method

BinaryConvRBM(visible_fields, hidden_fields, w)

Convolutional RBM with binary visible and hidden units, with fields a and b and weights w.

ConvolutionalRBMs.BinaryConvRBM — Method

BinaryConvRBM(visible_size, hidden_size, kernel_size)
BinaryConvRBM(T, visible_size, hidden_size, kernel_size)

Convolutional binary RBM with given dimensions, and parameters initialized to zero of type T (= Float64 by default).

ConvolutionalRBMs.conv_h2v — Method

conv_h2v(w, v)

Internal function used to compute inputs from a hidden configurations h to the visible layer, where w are the convolutional RBM weights.

\[I_{i_1,\dots,i_n}^{k_1,\dots,k_n,b} = \sum_{c,j_1,\dots,j_n} w_{c,j_1,\dots,j_n,\mu} v_{c,j_1+k_1-1,\dots,j_n+k_n-1,b}\]

Assumes that:

w is of size (C,J₁,...,Jₙ,M)

h is of size (M,K₁,...,Kₙ,B)

Here C is the channel dimension, M is the number of hidden units and B is the batch size. These three dimensions are flat, so that the hidden and batch dimensions must be flattened before calling conv_h2v. The output I is of size (C, K₁ + J₁ - 1, ..., Kₙ + Jₙ - 1, B).

Warning

This is an internal function and is not part of the public API.

Warning

Only works for n = 1, 2, 3, 4. Due to a technical limitation of Tullio.jl (#129), conv_h2v is defined by hand for different values of n. Therefore only a small number can be supported. If you need some higher value of n, consider doing a PR to define the corresponding conv_h2v method.

ConvolutionalRBMs.conv_v2h — Method

conv_v2h(w, v)

Internal function used to compute inputs from a visible configurations v to the hidden layer, where w are the convolutional RBM weights.

\[I_\mu^{k_1,\dots,k_n,b} = \sum_{c,j_1,\dots,j_n} w_{c,j_1,\dots,j_n,\mu} v_{c,j_1+k_1-1,\dots,j_n+k_n-1,b}\]

Assumes that:

w is of size (C,J₁,...,Jₙ,M)

v is of size (C,N₁,...,Nₙ,B)

Here C is the flattened channel dimension, M is the number of hidden units and B is the batch size. Therefore the hidden and batch dimensions must be flattened before calling conv_v2h. The output I is of size (M, N₁ - J₁ + 1, ..., Nₙ - Jₙ + 1, B).

Warning

This is an internal function and is not part of the public API.

Warning

Only works for n = 1, 2, 3 due to a technical limitations.

ConvolutionalRBMs.hsizes — Method

hsizes(convrbm, h) -> (hidden_size, output_size, batch_size)

Returns a (named) tuple decomposition of the size of h, such that:

size(h) == (hidden_size..., output_size..., batch_size...)

Throws an error if h is not consistent with this size.

ConvolutionalRBMs.input_size — Method

input_size(convrbm, h)

Input size of the visible configuration corresponding to h.

ConvolutionalRBMs.output_size — Method

output_size(convrbm, v)

Output size of the hidden configuration corresponding to v.

ConvolutionalRBMs.parts — Method

parts(denseconvrbm)

Extracts the dense and convolutional parts of a DenseConvRBM.

ConvolutionalRBMs.reshape_maybe — Method

reshape_maybe(x, shape)

Like reshape(x, shape), except that zero-dimensional outputs are returned as scalars.

ConvolutionalRBMs.vsizes — Method

vsizes(convrbm, v) -> (channel_size, input_size, batch_size)

Returns a (named) tuple decomposition of the size of v, such that:

size(v) == (channel_size..., input_size..., batch_size...)

Throws an error if v is not consistent with this size.

ConvolutionalRBMs.HankelTensors.Hankel — Method

Hankel(image, kernel_size, channel_size)

Constructs a Hankel array A from the underlying image data, such that:

A[c,j,k,b] = image[c,j+k-1,b]

where c,j,k,b are multi-indices for the channel, kernel, image, and batch dimensions. So j is the kernel entry, and k the kernel position in the image.

ConvolutionalRBMs.HankelTensors.hankel — Method

hankel(v, cdims, ksz)

Creates a Hankel array from v with channel dimensions cdims and kernel size ksz. The returned array A satisfies:

\[A_{i,j,k,n} = v_{i,j+k-1,n}\]

where i,j,k,n are multi-indices, with i traversing channels and j traversing the kernel. ```