Reference

ConvolutionalRBMs.ConvBinary — Type

ConvBinary(θ)

ConvolutionalRBMs._visconv — Function

_visconv(w, v)

Internal function used to compute inputs from a visible configuration v to the hidden layer, where w are the RBM weights.

\[I_\mu^{\mu,k_1,\dots,k_n,b} = \sum_{i_1,\dots,i_n} w_{i_1,\dots,i_n,\mu} v_{i_1+k_1-1,\dots,i_n+k_n-1,b}\]

Assumes that:

w is of size (N₁, N₂, ..., Nₙ, M)

v is of size (V₁, V₂, ..., Vₙ, B)

where M is the number of hidden units and B is the batch size. Therefore the hidden and batch dimension must be flattened before calling _visconv. The output I is of size (V₁ - N₁ + 1, V₁ - N₂ + 1, ..., Vₙ - Nₙ + 1, B).

Warning

This is an internal function and is not part of the public API. It can change in future (minor non-breaking) versions.

Warning

Only works for n = 1, 2, 3, 4. Due to a technical limitation of Tullio.jl (#129), _visconv is defined by hand for different values of n. Therefore only a small number can be supported. If you need some higher value of n, consider doing a PR to define the corresponding _visconv method.