Reference

ConvolutionalRBMs.ConvRBM — Method

ConvRBM(visible, hidden, weights; stride = 1, pad = 0, dilation = 1, groups = 1, pool = false)

Convolutional RBM.

pad == :same uses samepad(...) to infer a padding that preserves spatial dimensions.

ConvolutionalRBMs.ConvRBM — Method

ConvRBM(visible, hidden, kernel_size; kwargs...)

Convolutional RBM with given kernel_size.

ConvolutionalRBMs.BinaryConvRBM — Method

BinaryConvRBM(visible_fields, hidden_fields, w)

Convolutional RBM with binary visible and hidden units, with fields a and b and weights w.

ConvolutionalRBMs.BinaryConvRBM — Method

BinaryConvRBM(visible_size, hidden_size, kernel_size)
BinaryConvRBM(T, visible_size, hidden_size, kernel_size)

Convolutional binary RBM with given dimensions, and parameters initialized to zero of type T (= Float64 by default).

ConvolutionalRBMs.cartesian_sample_from_logits — Method

cartesian_sample_from_logits(logits; dims=1)

Returns an array X of indices sampled from CartesianIndices(logits) along dimensions dims, with probabilities P = softmax(logits; dims). In particular, dimensions dims of X are singleton.

ConvolutionalRBMs.conv_h2v — Method

conv_h2v(w, v)

Internal function used to compute inputs from a hidden configurations h to the visible layer, where w are the convolutional RBM weights.

\[I_{i_1,\dots,i_n}^{k_1,\dots,k_n,b} = \sum_{c,j_1,\dots,j_n} w_{c,j_1,\dots,j_n,\mu} v_{c,j_1+k_1-1,\dots,j_n+k_n-1,b}\]

Assumes that:

w is of size (C,J₁,...,Jₙ,M)

h is of size (M,K₁,...,Kₙ,B)

Here C is the channel dimension, M is the number of hidden units and B is the batch size. These three dimensions are flat, so that the hidden and batch dimensions must be flattened before calling conv_h2v. In the default case with stride = 1, pad = 0, dilation = 1, ..., the output I is of size (C, K₁ + J₁ - 1, ..., Kₙ + Jₙ - 1, B).

Warning

This is an internal function, not part of the public API. It is subject to change.

ConvolutionalRBMs.conv_v2h — Method

conv_v2h(w, v)

Internal function used to compute inputs from a visible configurations v to the hidden layer, where w are the convolutional RBM weights.

\[I_\mu^{k_1,\dots,k_n,b} = \sum_{c,j_1,\dots,j_n} w_{c,j_1,\dots,j_n,\mu} v_{c,j_1+k_1-1,\dots,j_n+k_n-1,b}\]

Assumes that:

w is of size (C,J₁,...,Jₙ,M)

v is of size (C,N₁,...,Nₙ,B)

Here C is the flattened channel dimension, M is the number of hidden units and B is the batch size. Therefore the hidden and batch dimensions must be flattened before calling conv_v2h. In the default case with stride = 1, pad = 0, dilation = 1, ..., the output I is of size (M, N₁ - J₁ + 1, ..., Nₙ - Jₙ + 1, B),

Warning

This is an internal function and is not part of the public API.

Warning

Only works for n = 1, 2, 3 due to a technical limitations.

ConvolutionalRBMs.expand_tuple — Method

expand_tuple(Val(N), t)

Expands t into a tuple of length N.

ConvolutionalRBMs.hankel — Method

hankel(convrbm, input_size)

Returns a dense RBM equivalent to convrbm with replicated weights.

ConvolutionalRBMs.hankel_image — Method

hankel_image(v, channel_size, kernel_size)

Creates a Hankel array from v with channel and kernel dimensions channel_size and kernel_size, respectively. The returned array V satisfies:

\[V_{c,j,k,n} = V_{c,j+k-1,n}\]

where c,j,k,n are multi-indices, with c traversing channels, j traversing the kernel, k the convolution output, and n batches.

ConvolutionalRBMs.hankel_weight — Method

hankel_weight(w, channel_size, input_size)

Creates a Hankel array from v with channel and kernel dimensions channel_size and kernel_size, respectively. The returned array V satisfies:

\[V_{c,j,k,n} = V_{c,j+k-1,n}\]

where c,j,k,n are multi-indices, with c traversing channels, j traversing the kernel, k the convolution output, and n batches.

ConvolutionalRBMs.hsizes — Method

hsizes(convrbm, h) -> (hidden_size, output_size, batch_size)

Returns a (named) tuple decomposition of the size of h, such that:

size(h) == (hidden_size..., output_size..., batch_size...)

Throws an error if h is not consistent with this size.

ConvolutionalRBMs.out2in — Method

out2in(j, k; stride = 1, dilation = 1)

Given kernel index j and output index k, gets the corresponding input index i.

ConvolutionalRBMs.output_size — Method

output_size(rbm, v)

Output size of the convolution.

If v has batch_size batches, then output_size(rbm, v) returns output_size such that:

size(h) == (hidden_size..., output_size..., batch_size...)

ConvolutionalRBMs.output_size — Method

output_size(kernel_size, input_size; stride = 1, pad = 0, dilation = 1)

Output size of the convolution.

ConvolutionalRBMs.parts — Method

parts(denseconvrbm)

Extracts the dense and convolutional parts of a DenseConvRBM.

ConvolutionalRBMs.randgumbel — Method

randgumbel(T = Float64)

Generates a random Gumbel variate.

ConvolutionalRBMs.replicate — Method

replicate(layer, n...)

Returns a new layer of size (size(layer)..., n...) by repeating the original layer along the new dimensions.

ConvolutionalRBMs.reshape_maybe — Method

reshape_maybe(x, shape)

Like reshape(x, shape), except that zero-dimensional outputs are returned as scalars.

ConvolutionalRBMs.samepad — Function

samepad(kernel_size, dilation = 1)

Computes padding such that the input and output spatial sizes of the convolution are the same.

If stride > 1, then size(output, d) * stride[d] == size(input, d) for each spatial dimension d.

ConvolutionalRBMs.splitpad — Method

splitpad(pad)

Given pad = (x_lo, x_hi, y_lo, y_hi, ...), returns lo = (x_lo, y_lo, ...), hi = (x_hi, y_hi, ...).

ConvolutionalRBMs.translate — Method

translate(v, channel_size, input_size, Δ; mode = :pad, fillvalue = 0)

Translates an image along input dimensions by Δ.

ConvolutionalRBMs.vsizes — Method

vsizes(convrbm, v) -> (channel_size, input_size, batch_size)

Returns a (named) tuple decomposition of the size of v, such that:

size(v) == (channel_size..., input_size..., batch_size...)

Throws an error if v is not consistent with this size.